fix(develop): implementacion de sneakerCard y correcciones de diseño responsive
