+++ aplicaciones_interactivas_tpo/src/ecomerce/components/Destacados.jsx
@@ -1,12 +1,10 @@
-import { Box, Button, Card, CardContent, CardMedia, Grid, Typography } from "@mui/material"
+import { Grid, Typography } from "@mui/material"
 import { dataDestacados } from '../data/dataDestacados'
-import { useCart } from "../../Cart/hooks/useCart"
 import { useLocation, useNavigate } from "react-router-dom";
 import { useEffect } from "react";
+import { SneakerCard } from "./SneakerCard";
 
 export const Destacados = () => {
-    const { addProduct } = useCart();
-    const navigate = useNavigate();
     const location = useLocation();
     useEffect(() => {
         if (location.hash) {
@@ -29,39 +27,8 @@ export const Destacados = () => {
             </Typography>
             
             <Grid container spacing={2} direction="row">
-                {dataDestacados.map(({ id, price, model, brand, colors, image }) => (
-                    <Grid key={id} size={{ xs: 12, sm: 6, md: 4, lg: 3 }}>
-                        <Card sx={{ cursor: 'pointer' }} onClick={() => navigate(`/producto/${id}`)}>
-                            <CardMedia component="img" alt={model} image={image[0]} />
-                            <CardContent sx={{ height: '200px' }}>
-                                <Box sx={{
-                                    display: 'inline-flex',
-                                    alignItems: 'center',
-                                    justifyContent: 'center',
-                                    backgroundColor: '#f65454',
-                                    border: '1px solid #ff7e7e',
-                                    padding: '2px 6px',
-                                    mb: 2
-                                }}>
-                                    <Typography sx={{
-                                        color: '#ffffff',
-                                        fontFamily: 'Inter',
-                                        fontWeight: 700,
-                                        fontStyle: 'italic',
-                                        fontSize: 12,
-                                    }}>
-                                        NEW IN
-                                    </Typography>
-                                </Box>
-
-                                <Typography gutterBottom variant="h5">${price}</Typography>
-                                <Typography gutterBottom variant="h6">{model}</Typography>
-                                <Typography variant="h6" sx={{ color: '#C0C0C0' }}>{brand}</Typography>
-                                <Typography variant="h6" sx={{ color: '#C0C0C0' }}>{colors.length} colores</Typography>
-                                
-                            </CardContent>
-                        </Card>
-                    </Grid>
+                {dataDestacados.map( sneaker => (
+                    <SneakerCard key={ sneaker.id } { ...sneaker } />
                 ))}
             </Grid>
         </>
